Anveshak/backend/requirements.txt
Flask==2.3.0
werkzeug==2.3.0
scapy==2.5.0
dpkt==1.9.8
pandas
numpy
matplotlib
reportlab==4.0.0
email-validator==2.0.0
scikit-learn
requests==2.31.0
google-generativeai
python-dotenv
